+++ operators.h
@@ -1,4 +1,4 @@
-double add(double op1, double op2);
-double mul(double op1, double op2);
-double div(double op1, double op2);
-double minus(double op1, double op2);
+double add(int op1, int op2);
+double mul(int op1, int op2);
+double div(int op1, int op2);
+double minus(int op1, int op2);
